feat: NormalFeedCellViewModel 생성
