--- burstcamp/burstcamp/Model/StateIndexPath.swift
@@ -1,13 +1,0 @@
-//
-//  StateIndexPath.swift
-//  burstcamp
-//
-//  Created by youtak on 2022/12/03.
-//
-
-import Foundation
-
-struct StateIndexPath {
-    var state: Bool
-    let indexPath: IndexPath?
-}
